+++ CHANGELOG.md
@@ -28,6 +28,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 
 - Set return statuses for SteamApp.update and BaseService.post_update as bool
+- Install git if pip source is github (useful in development/CI)
+- Bash install script updated to use return statuses and logging
 
 ### Deprecated